update location :)
update location :)

location & mv controller

2020 06 21
