+++ .settings/org.eclipse.wst.common.component
@@ -4,6 +4,7 @@
 		<wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
 		<wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
 		<wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
+        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/test/java"/>
 		<property name="java-output-path" value="target/classes"/>
         <property name="context-root" value="movie"/>
 	</wb-module>
