+++ src/main/java/com/korea/movie/BoardController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import common.Common;
-import common.Paging;
 import common.Paging2;
 import dao.BoardDAO;
 import vo.BoardVO;
@@ -23,7 +22,6 @@ import vo.UserVO;
 
 @Controller
 public class BoardController {
-
 
 	@Autowired
 	HttpServletRequest request;
@@ -34,128 +32,8 @@ public class BoardController {
 	public void setBoard_dao(BoardDAO board_dao) {
 		this.board_dao = board_dao;
 	}
-
-	//type2 : 현재상영작
-	@RequestMapping("/movieInfoDetailRank.do")
-	public String goMovieInfoDetail2(Model model, Integer page, String releaseDts, String title, String trailer) {
-		int type = 2;
-		int nowPage = 1;
-		if(page != null) {
-			nowPage = page;
-		}
-
-		//한 페이지에 표시되는 게시물의 시작과 끝 번호를 계산
-		int start = (nowPage - 1) * Common.Board.BLOCKLIST + 1;
-		int end = start + Common.Board.BLOCKLIST - 1;
-
-		Map<String, Object> map = new HashMap<String, Object>();
-		map.put("start", start);
-		map.put("end", end);
-		map.put("m_name", title);
-
-		List<BoardVO>list = null;
-		list = board_dao.selectList(map);
-
-		String content = "";
-		for(int i = 0; i < list.size(); i++) {
-			content = list.get(i).getContent().replaceAll("\n", "<br>");
-			list.get(i).setContent(content);
-		}
-
-		//전체 게시물 수 구하기
-		int row_total = board_dao.getRowTotal(title);
-
-		//Paging클래스를  사용하여 페이지 메뉴 생성하기                                                                               추가해야하나 "&m_name="+m_name,
-
-		String pageMenu = Paging.getPaging("movieInfoDetailRank.do?title="+title+"&releaseDts="+releaseDts, nowPage, row_total, Common.Board.BLOCKLIST, Common.Board.BLOCKPAGE);
-		int bunmo = board_dao.selectNum(title);
-		if(bunmo == 0) {
-			bunmo = 1;
-		}
-
-		float avg_f = (float)board_dao.selectSum(title) / bunmo;
-		float avg_f2 = Float.parseFloat(String.format("%.1f", avg_f));
-		int avg = board_dao.selectSum(title) / bunmo;
-
-		// String user_m_name = board_dao.selectM(title);
-
-		model.addAttribute("list", list);
-		model.addAttribute("avg_f2", avg_f2);
-		model.addAttribute("avg", avg);
-
-		// model.addAttribute("user_m_name", user_m_name);
-		model.addAttribute("pageMenu", pageMenu);
-		model.addAttribute("type", type);
-		model.addAttribute("releaseDts", releaseDts);
-		model.addAttribute("title", title);
-		model.addAttribute("trailer", trailer);
-		model.addAttribute("count", row_total);
-
-		return Common.Movie.VIEW_PATH + "movie_detail.jsp";
-	}
-
-	//영화별 전체 리뷰보기
-	//type1 : 상영예정작
-	@RequestMapping("/movieInfoDetail.do")
-	public String list(Model model, Integer page, String movieId, String movieSeq, String m_name) {
-		int type = 1;
-		int nowPage = 1;
-		if(page != null) {
-			nowPage = page;
-		}
-
-		//한 페이지에 표시되는 게시물의 시작과 끝 번호를 계산
-		int start = (nowPage - 1) * Common.Board.BLOCKLIST + 1;
-		int end = start + Common.Board.BLOCKLIST - 1;
-
-		Map<String, Object> map = new HashMap<String, Object>();
-		map.put("start", start);
-		map.put("end", end);
-		map.put("m_name", m_name);
-
-		List<BoardVO>list = null;
-		list = board_dao.selectList(map); 
-
-		String content = "";
-		for(int i = 0; i < list.size(); i++) {
-			content = list.get(i).getContent().replaceAll("\n", "<br>");
-			list.get(i).setContent(content);
-		}
-
-		//전체 게시물 수 구하기
-		int row_total = board_dao.getRowTotal(m_name); 
-
-		//Paging클래스를  사용하여 페이지 메뉴 생성하기
-		String pageMenu = Paging.getPaging("movieInfoDetail.do?movieId="+movieId +"&movieSeq="+movieSeq +"&m_name="+m_name, nowPage, row_total, Common.Board.BLOCKLIST, Common.Board.BLOCKPAGE);
-
-		int bunmo = board_dao.selectNum(m_name);
-		if(bunmo == 0) {
-			bunmo = 1;
-		}
-		int avg = board_dao.selectSum(m_name) / bunmo;
-
-		float avg_f = (float)board_dao.selectSum(m_name) / bunmo;
-		float avg_f2 = Float.parseFloat(String.format("%.1f", avg_f));
-		//String user_m_name = board_dao.selectM(m_name);
-
-
-		model.addAttribute("list", list);
-		model.addAttribute("avg_f2", avg_f2);
-		model.addAttribute("avg", avg);
-
-		// model.addAttribute("user_m_name", user_m_name);
-		model.addAttribute("pageMenu", pageMenu);
-		model.addAttribute("type", type);
-		model.addAttribute("m_name", m_name);
-		model.addAttribute("movieId", movieId);
-		model.addAttribute("movieSeq", movieSeq);
-		model.addAttribute("count", row_total);
-
-
-		return Common.Movie.VIEW_PATH + "movie_detail.jsp";
-	}
-
-	//로그인 되있는지 확인
+	
+	//로그인 확인하는 매핑
 	@RequestMapping("/checkLogin.do")
 	@ResponseBody
 	public String checkLogin(String id, String m_name) {//String id
@@ -197,20 +75,14 @@ public class BoardController {
 	//리뷰 등록
 	@RequestMapping("/insert.do")
 	public String insert(BoardVO vo, String type, String totalVar1, String totalVar2) {
-		System.out.println("리뷰 등록시 로그인되있는 아이디"+vo.getId() + "/영화이름 :" + vo.getM_name());
-		System.out.println("타입은 : "+type);
-		/*
-		 * String m_name = vo.getM_name().trim(); vo.setM_name(m_name);
-		 */
+
 		String content = vo.getContent().replaceAll("<br>", "\n");
 		vo.setContent(content);
 		board_dao.insert(vo);
 
 		if( type.equals("1") ) { 
-			System.out.println("등록타입1 지나가요");
 			return "redirect:movieInfoDetail.do&movieId="+totalVar1+"&movieSeq="+totalVar2+"&m_name="+vo.getM_name();
 		} else {
-			System.out.println("등록타입2 지나가요");
 			return "redirect:movieInfoDetailRank.do&title="+totalVar1+"&releaseDts="+totalVar2;
 		}
 	}
@@ -218,8 +90,6 @@ public class BoardController {
 	//수정 
 	@RequestMapping("/modify.do")
 	public String modify(BoardVO vo, String type, String totalVar1, String totalVar2) {
-		System.out.println("수정인데 로그인되있는 아이디"+vo.getId() + "/영화이름 :" + vo.getM_name());
-		System.out.println("타입은 : "+type);
 		String content = vo.getContent().replaceAll("<br>", "\n");
 		vo.setContent(content);
 
@@ -244,9 +114,7 @@ public class BoardController {
 			res = "yes";
 		}
 		return res;
-
 	}
-
 
 	@RequestMapping("/review.do")
 	public String reviews(Model model, Integer page) {
@@ -264,7 +132,6 @@ public class BoardController {
 		map.put("start", start);
 		map.put("end", end);
 
-
 		List<BoardVO>list = null;
 		list = board_dao.selectListTotal(map);
 
@@ -273,7 +140,6 @@ public class BoardController {
 			content = list.get(i).getContent().replaceAll("\n", "<br>");
 			list.get(i).setContent(content);
 		}
-
 
 		//전체 게시물 수 구하기
 		int row_total = board_dao.getRowTotal2();
@@ -284,10 +150,8 @@ public class BoardController {
 		model.addAttribute("pageMenu", pageMenu); 
 		model.addAttribute("count", row_total);
 
-
 		return Common.Board.VIEW_PATH + "community.jsp";   
 	}
-
 
 	@RequestMapping("/selectreview.do")
 	@ResponseBody
