+++ src/main/resources/config/spring/context/context-3-dao.xml
@@ -3,10 +3,6 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans https://www.springframework.org/schema/beans/spring-beans.xsd">
 	
-	<!-- <bean id="board_daoBean" class="dao.BoardDAO">
-		<property name="sqlSession" ref="sqlSessionBean"/> 
-	</bean> -->
-	
 	<bean id="board_daoBean" class="dao.BoardDAO">
 		<property name="sqlSession" ref="sqlSessionBean"/> 
 	</bean>
