+++ src/main/resources/config/spring/mvc/servlet-context.xml
@@ -26,12 +26,13 @@
 	<!-- 컨트롤러 수동 생성시 autowired사용 가능하게 하는 속성 -->
 	<context:annotation-config/>
 	
-	<!-- 수동생성 -->
+	<!-- MovieListController 수동생성 -->
 	<beans:bean class="com.korea.movie.MovieListController"> 
-		<beans:property name="movieListDAO" ref="movie_daoBean"/>   
+		<beans:property name="movieListDAO" ref="movie_daoBean"/>
+		<beans:property name="board_dao" ref="board_daoBean"/> 
 	</beans:bean>
 	
-	<!-- TicketController -->
+	<!-- TicketController 수동생성 -->
 	<beans:bean class="com.korea.movie.TicketController">
 			<beans:constructor-arg ref="theaterdao_Bean"/>
 			<beans:constructor-arg ref="ticketdao_Bean"/>
