+++ src/main/webapp/WEB-INF/views/movie/movie_detail.jsp
@@ -46,8 +46,6 @@
          }
       };
       
-      
-      
       //여러개 포스터 잘라쓰기
       function spPoster(inPosters) {
